requirements.txt
django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
PyMySQL==1.1.1
django-cors-headers==4.3.0
cryptography>=41.0.0

# gunicorn
setuptools
whitenoise
python-dotenv
# django-environ
# python-decouple